clean up lib and tests after replacing core igloo logic with new igloo/core package, retested.
