--- src/__tests__/bifrost/generateKeysetWithSecret.test.ts
@@ -1,195 +1,0 @@
-/**
- * Tests for the Bifrost generateKeysetWithSecret function
- */
-
-// Import the shared mock setup
-import { setupBuffMock } from '../__mocks__/buff.mock';
-setupBuffMock();
-
-// Define the types and mocks
-interface GroupPackage {
-  threshold: number;
-  commits: CommitPackage[];
-  group_pk: string;
-}
-
-interface CommitPackage {
-  idx: number;
-  pubkey: string;
-  hidden_pn: string;
-  binder_pn: string;
-}
-
-interface SharePackage {
-  idx: number;
-  binder_sn: string;
-  hidden_sn: string;
-  seckey: string;
-}
-
-// Create generateKeysetWithSecret mock
-const generateKeysetWithSecret = jest.fn().mockImplementation(
-  (threshold: number, totalMembers: number, secretKey: string) => {
-    // Validate input like the original function
-    if (threshold <= 0 || totalMembers <= 0) {
-      throw new Error('Threshold and total members must be positive numbers');
-    }
-    if (threshold > totalMembers) {
-      throw new Error('Threshold cannot be greater than total members');
-    }
-    if (!secretKey || typeof secretKey !== 'string') {
-      throw new Error('Secret key must be a non-empty string');
-    }
-    
-    // Add constraint for unreasonably large values
-    if (threshold > 100 || totalMembers > 100) {
-      throw new Error('Threshold and member count must be reasonable');
-    }
-    
-    // For nsec format, simulate decoding
-    let processedKey = secretKey;
-    if (secretKey.startsWith('nsec')) {
-      // Use a hash of the input key to simulate different outputs for different keys
-      processedKey = secretKey.split('').reduce((acc, char) => acc + char.charCodeAt(0), 0).toString();
-    }
-    
-    // Return a realistic result structure with outputs dependent on the inputs
-    const shareCredentials = Array(totalMembers).fill(0).map((_, i) => 
-      `mocked_share_pkg_${processedKey.substring(0, 8)}_${i+1}`
-    );
-    
-    return {
-      groupCredential: `mocked_group_pkg_${processedKey.substring(0, 8)}_${threshold}_of_${totalMembers}`,
-      shareCredentials
-    };
-  }
-);
-
-// Mock the module
-jest.mock('../../lib/bifrost', () => ({
-  generateKeysetWithSecret
-}));
-
-describe('Bifrost generateKeysetWithSecret Function', () => {
-  // Reset mocks before each test
-  beforeEach(() => {
-    generateKeysetWithSecret.mockClear();
-  });
-
-  it('should generate a keyset with valid parameters', () => {
-    const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-    const threshold = 2;
-    const totalMembers = 3;
-    
-    const result = generateKeysetWithSecret(threshold, totalMembers, secretKey);
-    
-    // Verify function was called with correct parameters
-    expect(generateKeysetWithSecret).toHaveBeenCalledWith(threshold, totalMembers, secretKey);
-    
-    // Verify the structure of the result
-    expect(result).toHaveProperty('groupCredential');
-    expect(result).toHaveProperty('shareCredentials');
-    expect(result.shareCredentials).toHaveLength(3);
-    
-    // Verify values reflect the parameters
-    expect(result.groupCredential).toContain('2_of_3');
-    expect(result.shareCredentials[0]).toContain('1');
-    expect(result.shareCredentials[1]).toContain('2');
-    expect(result.shareCredentials[2]).toContain('3');
-  });
-  
-  it('should handle nsec input format', () => {
-    const nsecKey = 'nsec1xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx';
-    const threshold = 2;
-    const totalMembers = 3;
-    
-    const result = generateKeysetWithSecret(threshold, totalMembers, nsecKey);
-    
-    // Verify function was called with correct parameters
-    expect(generateKeysetWithSecret).toHaveBeenCalledWith(threshold, totalMembers, nsecKey);
-    
-    // Verify the structure is the same as with hex key
-    expect(result).toHaveProperty('groupCredential');
-    expect(result).toHaveProperty('shareCredentials');
-    expect(result.shareCredentials).toHaveLength(3);
-  });
-  
-  it('should support different threshold configurations', () => {
-    // Test a 3-of-5 configuration
-    const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-    const threshold = 3;
-    const totalMembers = 5;
-    
-    const result = generateKeysetWithSecret(threshold, totalMembers, secretKey);
-    
-    expect(result.groupCredential).toContain('3_of_5');
-    expect(result.shareCredentials).toHaveLength(5);
-    
-    // Test a 1-of-1 configuration
-    const result2 = generateKeysetWithSecret(1, 1, secretKey);
-    
-    expect(result2.groupCredential).toContain('1_of_1');
-    expect(result2.shareCredentials).toHaveLength(1);
-  });
-  
-  it('should throw error when threshold is greater than total members', () => {
-    const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-    const threshold = 4;
-    const totalMembers = 3;
-    
-    expect(() => {
-      generateKeysetWithSecret(threshold, totalMembers, secretKey);
-    }).toThrow('Threshold cannot be greater than total members');
-  });
-  
-  it('should throw error when threshold or total members are invalid', () => {
-    const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-    
-    // Test negative threshold
-    expect(() => {
-      generateKeysetWithSecret(-1, 3, secretKey);
-    }).toThrow('Threshold and total members must be positive numbers');
-    
-    // Test zero threshold
-    expect(() => {
-      generateKeysetWithSecret(0, 3, secretKey);
-    }).toThrow('Threshold and total members must be positive numbers');
-    
-    // Test negative total members
-    expect(() => {
-      generateKeysetWithSecret(2, -1, secretKey);
-    }).toThrow('Threshold and total members must be positive numbers');
-  });
-  
-  it('should throw error when secret key is empty', () => {
-    const secretKey = '';
-    const threshold = 2;
-    const totalMembers = 3;
-    
-    expect(() => {
-      generateKeysetWithSecret(threshold, totalMembers, secretKey);
-    }).toThrow('Secret key must be a non-empty string');
-  });
-  
-  it('should throw error for very large threshold and member values', () => {
-    const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-    const threshold = 1000; // Unreasonably large threshold
-    const totalMembers = 2000; // Unreasonably large member count
-    
-    expect(() => {
-      generateKeysetWithSecret(threshold, totalMembers, secretKey);
-    }).toThrow('Threshold and member count must be reasonable');
-  });
-
-  it('should handle various nsec formats', () => {
-    // Test with different nsec formats
-    const validNsec1 = 'nsec1xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx';
-    const validNsec2 = 'nsec1yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy';
-    
-    // Both should work and produce different results
-    const result1 = generateKeysetWithSecret(2, 3, validNsec1);
-    const result2 = generateKeysetWithSecret(2, 3, validNsec2);
-    
-    expect(result1.groupCredential).not.toBe(result2.groupCredential);
-  });
-});
